ui 파일에서 json 싱크하게 설정
